requirements.txt
discord.py>=2.0.0
requests>=2.28.1
pdfplumber>=0.7.6
beautifulsoup4>=4.11.1
python-dotenv>=0.21.0
flask>=2.0.0